[MODIFY] 회원가입 에러처리 추가
