+++ src/main/java/tf/tailfriend/user/exception/PetTypeException.java
@@ -1,0 +1,20 @@
+package tf.tailfriend.user.exception;
+
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+import tf.tailfriend.global.exception.CustomException;
+import tf.tailfriend.user.message.error.ErrorMessage;
+
+@Getter
+public class PetTypeException extends CustomException {
+
+    @Override
+    public HttpStatus getStatus() {
+        return HttpStatus.BAD_REQUEST;
+    }
+
+    @Override
+    public String getMessage() {
+        return ErrorMessage.PETTYPE_NOTFOUND.getMessage();
+    }
+}
