+++ src/main/java/tf/tailfriend/user/message/error/ErrorMessage.java
@@ -1,0 +1,14 @@
+package tf.tailfriend.user.message.error;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum ErrorMessage {
+    USER_NOTFOUND("존재하지 않는 계정입니다"),
+    SNSTYPE_NOTFOUND("존재하지 않는 SNS 타입입니다"),
+    PETTYPE_NOTFOUND("존재하지 않는 PET 타입입니다");
+
+    private final String message;
+}
