+++ src/main/java/tf/tailfriend/user/service/UserService.java
@@ -1,119 +1,120 @@
-package tf.tailfriend.user.service;
-
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
-import tf.tailfriend.file.entity.File;
-import tf.tailfriend.file.service.FileService;
-import tf.tailfriend.global.service.StorageService;
-import tf.tailfriend.global.service.StorageServiceException;
-import tf.tailfriend.pet.entity.Pet;
-import tf.tailfriend.pet.entity.PetPhoto;
-import tf.tailfriend.pet.entity.PetType;
-import tf.tailfriend.pet.repository.PetPhotoDao;
-import tf.tailfriend.pet.repository.PetDao;
-import tf.tailfriend.pet.repository.PetTypeDao;
-import tf.tailfriend.user.entity.SnsType;
-import tf.tailfriend.user.entity.User;
-import tf.tailfriend.user.entity.dto.RegisterPetPhotoDto;
-import tf.tailfriend.user.entity.dto.RegisterPetDto;
-import tf.tailfriend.user.entity.dto.RegisterUserDto;
-import tf.tailfriend.user.repository.SnsTypeDao;
-import tf.tailfriend.user.repository.UserDao;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class UserService {
-
-    private final UserDao userDao;
-    private final PetDao petDao;
-    private final SnsTypeDao snsTypeDao;
-    private final PetTypeDao petTypeDao;
-    private final PetPhotoDao petPhotoDao;
-    private final FileService fileService;
-    private final StorageService storageService;
-
-    // ✅ 이메일로 userId 반환
-    public Integer getUserIdBySnsAccountId(String snsAccountId) {
-        return userDao.findBySnsAccountId(snsAccountId)
-                .map(User::getId)
-                .orElse(null);
-    }
-
-
-
-    @Transactional
-    public User registerUser(RegisterUserDto dto, List<MultipartFile> images) {
-        // 1. SNS 타입 조회
-        SnsType snsType = snsTypeDao.findById(dto.getSnsTypeId())
-                .orElseThrow(() -> new RuntimeException("SNS 타입 없음"));
-
-        // 2. 기본 프로필 파일
-        File defaultFile = fileService.getOrDefault(dto.getFileId());
-
-        // 3. 유저 저장
-        User user = User.builder()
-                .nickname(dto.getNickname())
-                .snsAccountId(dto.getSnsAccountId())
-                .snsType(snsType)
-                .file(defaultFile)
-                .build();
-
-        userDao.save(user);
-        userDao.flush(); // ✅ 즉시 DB 반영해서 user.id 보장
-
-        // 4. 펫 + 사진 등록
-        for (RegisterPetDto petDto : dto.getPets()) {
-            PetType petType = petTypeDao.findById(petDto.getPetTypeId())
-                    .orElseThrow(() -> new RuntimeException("펫 타입 없음"));
-
-            Pet pet = Pet.builder()
-                    .user(user)
-                    .petType(petType)
-                    .name(petDto.getName())
-                    .gender(petDto.getGender())
-                    .birth(petDto.getBirth())
-                    .weight(petDto.getWeight())
-                    .info(petDto.getInfo())
-                    .neutered(petDto.isNeutered())
-                    .activityStatus(petDto.getActivityStatus())
-                    .build();
-
-            petDao.save(pet);
-
-                int imageIndex = 0;
-
-                for (RegisterPetPhotoDto photoDto : petDto.getPhotos()) {
-                    if (imageIndex >= images.size()) break;
-
-                    MultipartFile image = images.get(imageIndex++);
-                    File file = fileService.save(image.getOriginalFilename(), "pet", photoDto.getType());
-
-                    try (InputStream is = image.getInputStream()) {
-                        storageService.upload(file.getPath(), is);
-                    } catch (IOException | StorageServiceException e) {
-                        throw new RuntimeException("파일 업로드 실패: " + e.getMessage(), e);
-                    }
-
-                PetPhoto petPhoto = PetPhoto.builder()
-                        .id(new PetPhoto.PetPhotoId(file.getId(), pet.getId()))
-                        .file(file)
-                        .pet(pet)
-                        .thumbnail(photoDto.isThumbnail())
-                        .build();
-
-                petPhotoDao.save(petPhoto);
-            }
-        }
-
-        return user;
-    }
-
-
-}
+//package tf.tailfriend.user.service;
+//
+//import jakarta.transaction.Transactional;
+//import lombok.RequiredArgsConstructor;
+//import org.springframework.stereotype.Service;
+//import org.springframework.web.multipart.MultipartFile;
+//import tf.tailfriend.file.entity.File;
+//import tf.tailfriend.file.service.FileService;
+//import tf.tailfriend.global.service.StorageService;
+//import tf.tailfriend.global.service.StorageServiceException;
+//import tf.tailfriend.pet.entity.Pet;
+//import tf.tailfriend.pet.entity.PetPhoto;
+//import tf.tailfriend.pet.entity.PetType;
+//import tf.tailfriend.pet.repository.PetPhotoDao;
+//import tf.tailfriend.pet.repository.PetDao;
+//import tf.tailfriend.pet.repository.PetTypeDao;
+//import tf.tailfriend.user.entity.SnsType;
+//import tf.tailfriend.user.entity.User;
+//import tf.tailfriend.user.entity.dto.RegisterPetPhotoDto;
+//import tf.tailfriend.user.entity.dto.RegisterPetDto;
+//import tf.tailfriend.user.entity.dto.RegisterUserDto;
+//import tf.tailfriend.user.exception.PetTypeException;
+//import tf.tailfriend.user.exception.SnsTypeException;
+//import tf.tailfriend.user.exception.UserException;
+//import tf.tailfriend.user.repository.SnsTypeDao;
+//import tf.tailfriend.user.repository.UserDao;
+//
+//import java.io.IOException;
+//import java.io.InputStream;
+//import java.util.List;
+//
+//@Service
+//@RequiredArgsConstructor
+//public class UserService {
+//
+//    private final UserDao userDao;
+//    private final PetDao petDao;
+//    private final SnsTypeDao snsTypeDao;
+//    private final PetTypeDao petTypeDao;
+//    private final PetPhotoDao petPhotoDao;
+//    private final FileService fileService;
+//    private final StorageService storageService;
+//
+//    // ✅ 이메일로 userId 반환
+//    public Integer getUserIdBySnsAccountId(String snsAccountId) {
+//        return userDao.findBySnsAccountId(snsAccountId)
+//                .map(User::getId)
+//                .orElseThrow(() -> new UserException());
+//    }
+//
+//
+//    @Transactional
+//    public User registerUser(RegisterUserDto dto, List<MultipartFile> images) {
+//        // 1. SNS 타입 조회
+//        SnsType snsType = snsTypeDao.findById(dto.getSnsTypeId())
+//                .orElseThrow(() -> new SnsTypeException());
+//
+//        // 2. 기본 프로필 파일
+//        File defaultFile = fileService.getOrDefault(dto.getFileId());
+//
+//        // 3. 유저 저장
+//        User user = User.builder()
+//                .nickname(dto.getNickname())
+//                .snsAccountId(dto.getSnsAccountId())
+//                .snsType(snsType)
+//                .file(defaultFile)
+//                .build();
+//
+//        userDao.save(user);
+//        userDao.flush(); // ✅ 즉시 DB 반영해서 user.id 보장
+//
+//        // 4. 펫 + 사진 등록
+//        for (RegisterPetDto petDto : dto.getPets()) {
+//            PetType petType = petTypeDao.findById(petDto.getPetTypeId())
+//                    .orElseThrow(() -> new PetTypeException());
+//
+//            Pet pet = Pet.builder()
+//                    .user(user)
+//                    .petType(petType)
+//                    .name(petDto.getName())
+//                    .gender(petDto.getGender())
+//                    .birth(petDto.getBirth())
+//                    .weight(petDto.getWeight())
+//                    .info(petDto.getInfo())
+//                    .neutered(petDto.isNeutered())
+//                    .activityStatus(petDto.getActivityStatus())
+//                    .build();
+//
+//            petDao.save(pet);
+//
+//                int imageIndex = 0;
+//
+//                for (RegisterPetPhotoDto photoDto : petDto.getPhotos()) {
+//                    if (imageIndex >= images.size()) break;
+//
+//                    MultipartFile image = images.get(imageIndex++);
+//                    File file = fileService.save(image.getOriginalFilename(), "pet", photoDto.getType());
+//
+//                    try (InputStream is = image.getInputStream()) {
+//                        storageService.upload(file.getPath(), is);
+//                    } catch (IOException | StorageServiceException e) {
+//                        throw new RuntimeException("파일 업로드 실패: " + e.getMessage(), e);
+//                    }
+//
+//                PetPhoto petPhoto = PetPhoto.builder()
+//                        .id(new PetPhoto.PetPhotoId(file.getId(), pet.getId()))
+//                        .file(file)
+//                        .pet(pet)
+//                        .thumbnail(photoDto.isThumbnail())
+//                        .build();
+//
+//                petPhotoDao.save(petPhoto);
+//            }
+//        }
+//
+//        return user;
+//    }
+//
+//
+//}
